pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "asap-to-kite"
version = "0.1.0"
description = "Convert ASAP-seq FASTQ files for processing with kite (kallisto | bustools)"
readme = "README.md"
authors = [
    {name = "Caleb Lareau"},
    {name = "Oliver Knight", email = "oliver.knight@charite.de"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "biopython>=1.80",
    "tqdm>=4.60",
]

[project.urls]
Repository = "https://github.com/ollieeknight/asap_to_kite"

[project.scripts]
asap-to-kite = "asap_to_kite.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["asap_to_kite*"]

[tool.black]
line-length = 100
target-version = ['py310']
skip-string-normalization = false

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
known_first_party = ["asap_to_kite"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "N", "PIE", "T20", "RET"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403", "F405"]
"cli.py" = ["T201"]  # print statements are intentional for --example flag
